api/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==38.0.1
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
drf-jwt==1.19.2
importlib-metadata==5.0.0
Markdown==3.4.1
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.3
pytz==2022.5
sqlparse==0.2.4
zipp==3.10.0
